kelvin/pyproject.toml
[project]
name = "kelvin"
version = "0.1.0"
dependencies = [
    "google-adk",
    "httpx",
    "google-auth",
    "litellm",
]

[tool.setuptools.packages.find]
where = ["./"]
include = ["text_agent"]
